+++ src/pbkdf2/pbkdf2_specific_sha1.h
@@ -43,7 +43,7 @@ void ps_sha1_append_int(pbkdf2_specific_sha1_ctx_t *ctx, uint32_t value);
 
 void ps_sha1_append_long(pbkdf2_specific_sha1_ctx_t *ctx, uint64_t value);
 
-void ps_sha1_append_str(pbkdf2_specific_sha1_ctx_t *ctx, unsigned char *str, uint32_t strlen);
+void ps_sha1_append_str(pbkdf2_specific_sha1_ctx_t *ctx, unsigned char *str, uint16_t strlen);
 
 uint32_t ps_rotate_left(uint32_t value, uint32_t shift);
 
@@ -51,7 +51,7 @@ uint32_t ps_rotate_right(uint32_t value, uint32_t shift);
 
 void ps_sha1(pbkdf2_specific_sha1_ctx_t *ctx);
 
-void ps_sha1_ctx_init(pbkdf2_specific_sha1_ctx_t *ctx, uint32_t num_of_chunks);
+void ps_sha1_ctx_init(pbkdf2_specific_sha1_ctx_t *ctx, uint8_t num_of_chunks);
 
 void ps_sha1_ctx_finalize(pbkdf2_specific_sha1_ctx_t *ctx);
